refactor: DoneContentView 삭제하고 3가지 콜렉션뷰가 CollectionContentView로 통일할수있도록 수정
